+++ src/routes/alphabet/group-[group]-[noise]-quiz/+page.js
@@ -1,0 +1,8 @@
+export async function load({ params }) {
+let noise = params.noise
+let group = params.group
+return {
+noise: noise,
+group: group
+}
+}
