+++ src/routes/books/[book_title]/[page_number]/+page.svelte
@@ -1,24 +1,14 @@
-<script context="module">
-    export async function load({ params }) {
-        let book_title = params.book_title
-        let page_number = params.page_number
-        return { props: {
-            book_title: book_title,
-            page_number: page_number
-        }}
-    }
-</script>
-
 <script>
     import books from "$lib/data/books.json"
     import PersonSpeakingBook from "$lib/components/buttons/person_speaking_book.svelte";
     import White from '$lib/components/buttons/styles/white_button.svelte'
 
-    export let book_title
-    export let page_number
+    export let data
+    $: page_number = data.page_number
+    $: book_title = data.book_title
+
     $: next_page = parseInt(page_number) + 1
     $: last_page = parseInt(page_number) - 1
-
     $: image_source = ['/books/', book_title, '/', page_number, '.jpeg'].join('')
     $: text = books[ book_title ][ page_number-1 ]
