+++ src/routes/letter/[letter]/+page.js
@@ -1,0 +1,5 @@
+export async function load({ params }) {
+let letter = params.letter
+
+return {letter}
+}
